store refactor
