+++ frontend/src/main.js
@@ -1,7 +1,7 @@
 import Vue from "vue";
 import App from "./App.vue";
 import vuetify from "./plugins/vuetify";
-import store from "./store/store";
+import store from "./store";
 import VueRouter from "vue-router";
 import { routes } from "./routes";
 import i18n from "./i18n";
