+++ frontend/src/utils/validators.js
@@ -1,0 +1,8 @@
+const validators = {
+  emailRules: v =>
+    !v ||
+    /^\w+([.-]?\w+)*@\w+([.-]?\w+)*(\.\w{2,3})+$/.test(v) ||
+    "E-mail must be valid",
+};
+
+export { validators };
